+++ src/padrao/dados/navegacao.js
@@ -1,7 +1,7 @@
 export const navegacao = [
-  { nome: "HOMEM", link: "" },
-  { nome: "MULHER", link: "" },
-  { nome: "GUIA DE EQUIPAMENTO", link: "" },
-  { nome: "ONDE ENCONTRAR", link: "" },
-  { nome: "INSTITUCIONAL", link: "" },
+  { nome: "HOMEM", link: "/homem" },
+  { nome: "MULHER", link: "/mulher" },
+  { nome: "GUIA DE EQUIPAMENTO", link: "/guia-de-equipamento" },
+  { nome: "ONDE ENCONTRAR", link: "/onde-encontrar" },
+  { nome: "INSTITUCIONAL", link: "/institucional" },
 ];